+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - refactor videojs xapi handling as videojs plugin
 - upgrade react-query to @tanstack/react-query (v4) (#2340)
 - Replace component SortableTable by Cunningham DataGrid (#2311)
+- Replace grommet DatePicker by Cunningham DatePicker (#2359)
 
 ### Fixed